API interfacing is working, basic metrics complete.

+++ app/models/data_point.rb
@@ -2,7 +2,7 @@ class DataPoint
   DataPointKey = 'hash_data'
 
   def self.range(start, finish)
-    $redis.zrangebyscore(DataPointKey, start.to_i, finish.to_i, with_scores: true).map { |a| {time: a[1].to_i, value: a[0].to_i} }
+    $redis.zrangebyscore(DataPointKey, start.to_i, finish.to_i, with_scores: true).map { |a| {time: a[1].to_i, value: a[0].to_f} }
   end
 
   def self.add(date, value)
